% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{coefs_to_points}
\alias{coefs_to_points}
\title{Map a set of elastic net logistic regression coeficients 1-1 to points where 100 is the chosen threshold}
\usage{
coefs_to_points(objModel, thresholdStrat = 0.4, pointsStrat = 100)
}
\arguments{
\item{objModel}{an elastic net logistic model fit using caret}

\item{thresholdStrat}{the percentage probability to cut off as positive}

\item{pointsStrat}{the threshold in points to map the cut off to}
}
\description{
Map a set of elastic net logistic regression coeficients 1-1 to points where 100 is the chosen threshold
}
