% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{create_int_lms}
\alias{create_int_lms}
\title{Take a dataframe and a set of variables and fit three logitic regressions using forward selection}
\usage{
create_int_lms(datfr, resp_var = "response", pred_vars = NA,
  int_var = NA, run_forward_selection = FALSE, trace_fs = FALSE,
  steps_fs = 5)
}
\arguments{
\item{datfr}{}

\item{resp_var}{}

\item{pred_vars}{}

\item{int_var}{}

\item{run_forward_selection}{}

\item{trace_fs}{}

\item{steps_fs}{}
}
\description{
Take a dataframe and a set of variables and fit three logitic regressions using forward selection
}
