% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{dbQueryReconnect}
\alias{dbQueryReconnect}
\title{Connect to DB, and if it's timed out, tries again}
\usage{
dbQueryReconnect(statement, con = conn, ...,
  conn_specs = dbConnect_specs, db_func = c("get", "send"))
}
\arguments{
\item{statement}{SQL code to query}

\item{conn_specs}{a list that has item 1 = drv, item 2 = host, item 3 = keyring service name that is the server with password of user password for server}

\item{db_func}{specifies which dbXQuery function to use, where X is the name of the db_func}

\item{conn}{an open DBIConnection object, as returned by dbConnect()}
}
\value{
data_frame object
}
\description{
Connect to DB, and if it's timed out, tries again
}
\examples{
dbQueryReconnect("SELECT * FROM GENDER_D")
dbQueryReconnect("DROP TABLE ABC_XYZ_123___", db_func = "send")
}
