% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_column_names_somewhere_in_data_frame}
\alias{expect_column_names_somewhere_in_data_frame}
\title{Check if the column names you expect to be in the df, are indeed in there}
\usage{
expect_column_names_somewhere_in_data_frame(df, colums_expected)
}
\arguments{
\item{colums_expected}{a character vector}
}
\description{
Check if the column names you expect to be in the df, are indeed in there
}
\examples{
expect_column_names_somewhere_in_data_frame(mtcars, c("mpg", "cyl"))
# [1] "all columns found...OK"
expect_column_names_somewhere_in_data_frame(mtcars, c("mpg", "cyl", "car_name"))
# Error in expect_column_names_somewhere_in_data_frame(mtcars, c("mpg",  :
#   car_name column not found
}
