% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{expect_no_duplicates}
\alias{expect_no_duplicates}
\title{Check if a data frame has duplicates in a given column. If a vector is given, check for duplicates in the vector.}
\usage{
expect_no_duplicates(df, group_by_column, stop_if_surprise = TRUE,
  report_duplicates = TRUE)
}
\arguments{
\item{df}{data frame to check}

\item{group_by_column}{character vector name of column expecting no duplicates}

\item{stop_if_surprise}{T/F for whether to consider failure an error}

\item{report_duplicates}{T/F for whether to return a partial list of the top duplicates if failure}
}
\value{
several options depending on whether it fails or succeeeds
}
\description{
Check if a data frame has duplicates in a given column. If a vector is given, check for duplicates in the vector.
}
\examples{
expect_no_duplicates(mtcars, "wt")
expect_no_duplicates(mtcars, "cyl")
}
