% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{in_sql_list}
\alias{in_sql_list}
\title{Syntatic sugar to turn a vector into a SQL "IN (x1, x2, x3, ...)" phrase}
\usage{
in_sql_list(x, nbrs = FALSE)
}
\arguments{
\item{x}{vector of values to search in}

\item{nbrs}{False if characters needing quotes, else will not quote values}
}
\value{
character vector object
}
\description{
Syntatic sugar to turn a vector into a SQL "IN (x1, x2, x3, ...)" phrase
}
\examples{
in_sql_list(mtcars$hp, nbrs = TRUE)
}
