% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_queries.R
\name{join_list_sql}
\alias{join_list_sql}
\title{Gets around limit of 1000 items in a list by iterating a query}
\usage{
join_list_sql(sortie_df, sql_left_join_to_in, sql_query_before_new_table,
  sql_query_after_new_table, conn)
}
\arguments{
\item{sortie_df}{data frame of one column potentially > 1000 rows needing iteration in SQL because it's too long to turn into a list}

\item{sql_left_join_to_in}{SQL code to SELECT the table to save as a left join starting point through the word IN just before the the (list)}

\item{sql_query_before_new_table}{SQL code to select table to return from SELECT up to and not including the name of the temp table}

\item{sql_query_after_new_table}{SQL code to select table to return after the name of the temp table through to the JOIN ON clause, etc.}

\item{conn}{an open DBIConnection object, as returned by dbConnect()}
}
\value{
data_frame object
}
\description{
Gets around limit of 1000 items in a list by iterating a query
}
\examples{
join_list_sql(mtcars, "wt", "mpg", "weight", "miles per gallon")
}
