% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{plot_predict_var}
\alias{plot_predict_var}
\title{Buids a variable profile plot with optional cutoffs at a vector that differentiated categorical mappings}
\usage{
plot_predict_var(x, y, xlabel = "x", ylabel = "y", cutoffs = NA)
}
\arguments{
\item{x}{predictor variable}

\item{y}{response variable}

\item{xlabel}{what to call your predictor}

\item{ylabel}{what to call your response}

\item{cutoffs}{an optional vector of cut points where a numeric variable becomes a continuous variable}
}
\value{
ggplot2 object
}
\description{
Buids a variable profile plot with optional cutoffs at a vector that differentiated categorical mappings
}
\examples{
plot_predict_var(mtcars$wt, mtcars$mpg, "weight", "miles per gallon")
plot_predict_var(mtcars$hp, mtcars$mpg, "horsepower", "miles per gallon", cutoffs = c(80, 135, 200))
}
