% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{train_model}
\alias{train_model}
\title{Train a ridge regression on a set of variables and thranslate coeficients into points}
\usage{
train_model(x = model_input, vars = input_vars,
  resp = "response_label", cont = objControl, thresholdStrat = 0.4,
  pointsStrat = 100)
}
\arguments{
\item{x}{}

\item{vars}{}

\item{resp}{}

\item{cont}{}

\item{thresholdStrat}{}
}
\description{
Train a ridge regression on a set of variables and thranslate coeficients into points
}
