% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_training.R
\name{translate_coefs_to_points}
\alias{translate_coefs_to_points}
\title{Map a generic logisitc model's coeficients 1-1 to points where 100 is stratification, handling relevelling of factors}
\usage{
translate_coefs_to_points(model_coef, thresholdStrat = 0.4,
  pointsStrat = 100)
}
\arguments{
\item{model_coef}{}

\item{thresholdStrat}{}

\item{pointsStrat}{}
}
\description{
Map a generic logisitc model's coeficients 1-1 to points where 100 is stratification, handling relevelling of factors
}
