% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_expectations.R
\name{wc}
\alias{wc}
\title{Write a data frame to a temporary .csv and open in Excel}
\usage{
wc(df, n = NA, open = TRUE, del = TRUE, del_wait = 10)
}
\arguments{
\item{n}{suffix for temporary csv file name, defaults to Sys.time()}

\item{open}{T/F for whether to auto-open in Excel}

\item{del}{T/F whether to deleted the created file}

\item{del_wait}{how many seconds to wait to delete the file}
}
\description{
Write a data frame to a temporary .csv and open in Excel
}
\examples{
wc(mtcars)
mtcars \%>\% filter(cyl == 4) \%>\% wc()
}
