% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{wssplot}
\alias{wssplot}
\title{Clustering Screeplot to help determine k}
\usage{
wssplot(df, nc = 15, seed = 1234, nstart = 1)
}
\arguments{
\item{df}{}

\item{nc}{}

\item{seed}{}

\item{nstart}{}
}
\description{
Clustering Screeplot to help determine k
}
